% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_kegg_db.R
\name{create_kegg_db_microbiota}
\alias{create_kegg_db_microbiota}
\title{create_kegg_db_microbiota}
\usage{
create_kegg_db_microbiota(
  output_dir = ".",
  include = c("bacteria", "archaea", "fungi", "viruses"),
  author = NULL,
  maintainer = NULL
)
}
\arguments{
\item{output_dir}{output directory to write the generated KEGG.db package}

\item{include}{microbiota groups to include}

\item{author}{a string for the Author field in the generated package DESCRIPTION}

\item{maintainer}{a string for the Maintainer field in the generated package DESCRIPTION}
}
\value{
paths of the generated package tarballs
}
\description{
Create KEGG.db package for microbiota
}
\details{
This function queries KEGG organism list and builds KEGG.db for selected microbiota groups.
}
