% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_kegg_db.R
\name{create_kegg_db}
\alias{create_kegg_db}
\title{create_kegg_db}
\usage{
create_kegg_db(species)
}
\arguments{
\item{species}{one of KEGG supported species, e.g. hsa for human}
}
\value{
KEGG.db package generated in working directory
}
\description{
create KEGG.db package
}
\author{
Guangchuang Yu and Ziru Chen
}
